requirements.txt
# Existing dependencies
crewai[openai,google-genai]==1.3.0
streamlit>=1.32.0
python-dotenv>=1.1.1
pandas>=2.2.0
requests>=2.31.0
openai>=1.51.0
tenacity>=8.2.0

# ===================================
# WEB SCRAPING & EXTRACTION
# ===================================
beautifulsoup4>=4.12.0
lxml>=5.1.0
html5lib>=1.1
selenium>=4.15.0
playwright>=1.42.0
scrapy>=2.11.0

# ===================================
# DATA EXTRACTION & PARSING
# ===================================
trafilatura>=1.6.3          # Clean article extraction
newspaper3k>=0.2.8          # News article parsing
scholarly>=1.7.11           # Google Scholar scraping
biopython>=1.83             # Biological data parsing
rdkit>=2023.9.1             # Chemistry data processing

# ===================================
# ENHANCED NLP (optional; moved to requirements-optional-nlp.txt)
# Install separately if needed and if your Python version has compatible wheels
# See requirements-optional-nlp.txt for pinned versions

# ===================================
# PDF & DOCUMENT PROCESSING
# ===================================
pypdf>=4.0.0
pdfplumber>=0.10.3
python-docx>=1.1.0